[Interjections] - Move interjections logic to a class.
